INTPYTHON-447 Ensure index is ready for retriever tests

+++ libs/mongodb/tests/integration_tests/test_retrievers.py
@@ -1,4 +1,5 @@
 import os
+from time import monotonic, sleep
 from typing import Generator, List
 
 import pytest
@@ -97,7 +98,7 @@ def indexed_vectorstore(
 ) -> Generator[MongoDBAtlasVectorSearch, None, None]:
     """Return a VectorStore with example document embeddings indexed."""
 
-    vectorstore = PatchedMongoDBAtlasVectorSearch(
+    vectorstore = MongoDBAtlasVectorSearch(
         collection=collection,
         embedding=embedding_openai,
         index_name=VECTOR_INDEX_NAME,
@@ -106,12 +107,28 @@ def indexed_vectorstore(
 
     vectorstore.add_documents(example_documents)
 
+    n_docs = collection.count_documents({})
+
+    start = monotonic()
+    found = False
+    while monotonic() - start <= TIMEOUT:
+        if (
+            len(vectorstore.similarity_search("sandwich", k=n_docs, oversampling_factor=1))
+            == n_docs
+        ):
+            found = True
+            break
+        else:
+            sleep(INTERVAL)
+    if not found:
+        raise TimeoutError("Failed to update index")
+
     yield vectorstore
 
     vectorstore.collection.delete_many({})
 
 
-def test_vector_retriever(indexed_vectorstore: PatchedMongoDBAtlasVectorSearch) -> None:
+def test_vector_retriever(indexed_vectorstore: MongoDBAtlasVectorSearch) -> None:
     """Test VectorStoreRetriever"""
     retriever = indexed_vectorstore.as_retriever()
 
@@ -125,7 +142,7 @@ def test_vector_retriever(indexed_vectorstore: PatchedMongoDBAtlasVectorSearch) 
     assert "New Orleans" in results[0].page_content
 
 
-def test_hybrid_retriever(indexed_vectorstore: PatchedMongoDBAtlasVectorSearch) -> None:
+def test_hybrid_retriever(indexed_vectorstore: MongoDBAtlasVectorSearch) -> None:
     """Test basic usage of MongoDBAtlasHybridSearchRetriever"""
     retriever = MongoDBAtlasHybridSearchRetriever(
         vectorstore=indexed_vectorstore,
@@ -144,7 +161,7 @@ def test_hybrid_retriever(indexed_vectorstore: PatchedMongoDBAtlasVectorSearch) 
 
 
 def test_fulltext_retriever(
-    indexed_vectorstore: PatchedMongoDBAtlasVectorSearch,
+    indexed_vectorstore: MongoDBAtlasVectorSearch,
 ) -> None:
     """Test result of performing fulltext search.
